requirements.txt
Flask==3.0.2
Flask_JWT_Extended==4.6.0
flask_sqlalchemy==3.1.1
pytest==8.0.0
Requests==2.31.0
SQLAlchemy==2.0.25
Werkzeug==3.0.1
python-dotenv==1.0.1